Linter rules to prefer const
We have lots of code that uses `var`, some code that uses `let` and almost no code that uses `const`. The latter is best in terms of reliability and 95% of the time it's what people should be using. This PR introduces 3 rules that should encourage the correct use of const, disallow the use of var and discourage using let when const would suffice.